Exctracted test helper classes
