+++ tests/Console/Commands/DiscoveryClearCommandTest.php
@@ -4,10 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Tempest\Console\Commands;
 
-use ReflectionClass;
 use Tempest\AppConfig;
-use Tempest\Container\Container;
-use Tempest\Discovery\Discovery;
 use Tests\Tempest\TestCase;
 
 class DiscoveryClearCommandTest extends TestCase
@@ -24,35 +21,5 @@ class DiscoveryClearCommandTest extends TestCase
         $this->console('discovery:clear');
 
         $this->assertTrue(MyDiscovery::$cacheCleared);
-    }
-}
-
-class MyDiscovery implements Discovery
-{
-    public static bool $cacheCleared = false;
-
-    public function discover(ReflectionClass $class): void
-    {
-        // TODO: Implement discover() method.
-    }
-
-    public function hasCache(): bool
-    {
-        return false;
-    }
-
-    public function storeCache(): void
-    {
-        // TODO: Implement storeCache() method.
-    }
-
-    public function restoreCache(Container $container): void
-    {
-        // TODO: Implement restoreCache() method.
-    }
-
-    public function destroyCache(): void
-    {
-        self::$cacheCleared = true;
     }
 }
