+++ tests/Container/BuiltinArrayClass.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class BuiltinArrayClass
+{
+	public function __construct(public array $anArray)
+	{
+	}
+}
