+++ tests/Container/BuiltinTypesWithDefaultsClass.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class BuiltinTypesWithDefaultsClass
+{
+	public function __construct(
+		public string $aString = 'This is a default value',
+	)
+	{
+	}
+}
