+++ tests/Container/CallContainerObjectE.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class CallContainerObjectE
+{
+	public function method(ContainerObjectE $input): ContainerObjectE
+	{
+		return $input;
+	}
+}
