+++ tests/Container/ContainerObjectA.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class ContainerObjectA
+{
+}
