+++ tests/Container/ContainerObjectB.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class ContainerObjectB
+{
+	public function __construct(public ContainerObjectA $a)
+	{
+	}
+}
