+++ tests/Container/ContainerObjectC.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class ContainerObjectC
+{
+	public function __construct(public string $prop)
+	{
+	}
+}
