+++ tests/Container/ContainerObjectD.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+use Tempest\Container\InitializedBy;
+
+#[InitializedBy(ContainerObjectDInitializer::class)]
+class ContainerObjectD
+{
+	public function __construct(public string $prop)
+	{
+	}
+}
