+++ tests/Container/ContainerObjectE.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class ContainerObjectE
+{
+	public function __construct(public string $id = 'default')
+	{
+	}
+}
