+++ tests/Container/OptionalTypesClass.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class OptionalTypesClass
+{
+	public function __construct(
+		public ?string $aString
+	)
+	{
+	}
+}
