+++ tests/Container/SingletonClass.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+class SingletonClass
+{
+	public static int $count = 0;
+
+	public function __construct()
+	{
+		self::$count += 1;
+	}
+}
