+++ tests/Container/UnionTypesClass.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Tests\Tempest\Container;
+
+use DateTime;
+
+class UnionTypesClass
+{
+	public function __construct(
+		public DateTime $aStringOrDate
+	)
+	{
+	}
+}
