+++ tests/ORM/Foo.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Tempest\ORM;
+
+use Tempest\ORM\IsModel;
+use Tempest\ORM\Model;
+
+class Foo implements Model
+{
+	use IsModel;
+
+	public string $bar;
+}
